# Conflicts:
                       #	config/adminlte.php

+++ routes/web.php
@@ -23,16 +23,28 @@ Route::get('/test', function () {
 });
 
 
-Route::get('region/', 'RegionController@index');
-Route::get('region/', 'RegionController@index')->name('region.index');
-Route::get('region/create', 'RegionController@create')->name('region.create');
-Route::get('region/createCity', 'RegionController@createCity')->name('region.createCity');
-Route::get('region/createDistrict', 'RegionController@createDistrict')->name('region.createDistrict');
-Route::get('region/findCity/{id}', 'RegionController@findCity');
-Route::get('region/edit/{id}', 'RegionController@edit')->name('region.edit');
-Route::post('region/', 'RegionController@store')->name('region.store');
-Route::patch('region/show/{id}', 'RegionController@update')->name('region.update');
-Route::delete('region/{id}', 'RegionController@destroy')->name('region.destroy');
+//Маршруты для Праздников
+Route::get('celebration/', 'Admin\CelebrationController@index');
+Route::get('celebration/', 'Admin\CelebrationController@index')->name('celebration.index');
+Route::get('celebration/create', 'Admin\CelebrationController@create')->name('celebration.create');
+Route::post('celebration/', 'Admin\CelebrationController@store')->name('celebration.store');
+Route::get('celebration/edit/{id}', 'Admin\CelebrationController@edit')->name('celebration.edit');
+Route::patch('celebration/show/{id}', 'Admin\CelebrationController@update')->name('celebration.update');
+Route::delete('celebration/{id}', 'Admin\CelebrationController@destroy')->name('celebration.destroy');
+
+//Маршруты для Регионов
+Route::get('region/', 'Admin\RegionController@index');
+Route::get('region/', 'Admin\RegionController@index')->name('region.index');
+Route::get('region/create', 'Admin\RegionController@create')->name('region.create');
+Route::get('region/createCity', 'Admin\RegionController@createCity')->name('region.createCity');
+Route::get('region/createDistrict', 'Admin\RegionController@createDistrict')->name('region.createDistrict');
+Route::get('region/findCity/{id}', 'Admin\RegionController@findCity');
+Route::get('region/edit/{id}', 'Admin\RegionController@edit')->name('region.edit');
+Route::get('region/editCity/{id}', 'Admin\RegionController@editCity')->name('region.editCity');
+Route::get('region/editDistrict/{id}', 'Admin\RegionController@editDistrict')->name('region.editDistrict');
+Route::post('region/', 'Admin\RegionController@store')->name('region.store');
+Route::patch('region/show/{id}', 'Admin\RegionController@update')->name('region.update');
+Route::delete('region/{id}', 'Admin\RegionController@destroy')->name('region.destroy');
 
 
 Auth::routes(['verify' => true]);
